feature: PRODUCT Integer > Long

+++ pefacil-product/pefacil-product-domain/src/main/java/br/com/pefacil/product/domain/port/api/DeleteProduct.java
@@ -2,5 +2,5 @@ package br.com.pefacil.product.domain.port.api;
 
 public interface DeleteProduct {
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
